Create gbff file parsing
here i have taken the example of a fungi refseq 
to extract out the accession number, the gene location, first amino acid and the first three nucleotide. 
the major goal of this parsing is to find out the exception in the genome of fungi species that have ATG as their start codon but do not code for  methionine or vice versa.. 
this is a general code using regex. 
we can also use BIOPYTHON functions to parse the file.